requirements.txt
Flask==1.0.2
pymongo==3.7.2
gunicorn==19.9.0
python-ldap==3.1.0
